+++ RTWWebServer/DTO/Request/CreateGuestRequest.cs
@@ -1,0 +1,5 @@
+namespace RTWWebServer.DTO.Request;
+
+public class CreateGuestRequest
+{
+}
