+++ RTWWebServer/DTO/response/GuestLoginResponse.cs
@@ -1,6 +1,6 @@
 using NetworkDefinition.ErrorCode;
 
-namespace RTWWebServer.RequestResponse;
+namespace RTWWebServer.DTO.response;
 
 public class GuestLoginResponse(WebServerErrorCode errorCode, string authToken) : IResponse
 {
