--- RTWWebServer/RequestResponse/CreateGuestRequest.cs
@@ -1,5 +1,0 @@
-namespace RTWWebServer.RequestResponse;
-
-public class CreateGuestRequest
-{
-}
